Added precision.

+++ include/pi.h
@@ -22,7 +22,7 @@
  *
  *  returns: the dth hex digit of pi (plus num_places - 1 more places)
  */
-char* pi_hex(int d, int num_places);
+char* pi_hex(long long int d, int num_places);
 
 /*
  * Function:  pi
@@ -34,7 +34,7 @@ char* pi_hex(int d, int num_places);
  *
  *  returns: pi starting at digit d
  */
-long double pi(int d);
+long double pi(long long int d);
 
 /*
  * Function:  pi_hex_p1 
@@ -89,7 +89,7 @@ long long int pi_hex_p4(long long int k);
  *
  *  returns: the first term of the pi hex BBP formula
  */
-long double compute_pi_sum1(int d);
+long double compute_pi_sum1(long long int d);
 
 /*
  * Function:  compute_pi_sum2 
@@ -100,7 +100,7 @@ long double compute_pi_sum1(int d);
  *
  *  returns: the second term of the pi hex BBP formula
  */
-long double compute_pi_sum2(int d);
+long double compute_pi_sum2(long long int d);
 
 /*
  * Function:  compute_pi_sum3 
@@ -111,7 +111,7 @@ long double compute_pi_sum2(int d);
  *
  *  returns: the third term of the pi hex BBP formula
  */
-long double compute_pi_sum3(int d);
+long double compute_pi_sum3(long long int d);
 
 /*
  * Function:  compute_pi_sum4 
@@ -122,7 +122,7 @@ long double compute_pi_sum3(int d);
  *
  *  returns: the fourth term of the pi hex BBP formula
  */
-long double compute_pi_sum4(int d);
+long double compute_pi_sum4(long long int d);
 
 #endif // PI_H
